Add missing flagEndAsync in acknowledge implementation
It was probably forgotten by the original author. Without it, when I've
completed one purchase flow, I have to quit and restart the app before I
can complete another. This isn't a big deal, since most apps are content
selling one subscription. But if the ‘Subscribe’ button stays active and
the user presses it again, the error message with flagEndAsync will be
more useful than without. It will tell that the subscription is already
active and offer a link to the subscription management page, rather than
showing a strange error about async operations.

I got the inspiration from
https://github.com/budiselic/cordova-plugin-inapppurchase-fixed/issues/1#issuecomment-1002835542
and
https://github.com/Filavision/cordova-plugin-inapppurchase-2/commit/0563d06af4876cff47df7d99cdf3b27b37fdabed
, but ended up putting flagEndAsync in a place that makes more sense.